fix: isolate client-owned web fetch
