+++ .gitattributes
@@ -1,2 +1,1 @@
 * text eol=lf
-*.png binary
